requirements.txt
ant @ git+https://github.com/mch/python-ant.git@87f395f57e1112ec1932eba8ddc5a0fd0b25f311
asgiref==3.4.1
certifi==2021.10.8
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
msgpack-python==0.5.6
pyserial==3.5
pytz==2021.3
pyusb==1.2.1
pywin32==303
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
wincertstore==0.2




